refactor: enum response 패키지 이동
