+++ src/main/java/com/poortorich/iteration/service/IterationService.java
@@ -3,7 +3,7 @@ package com.poortorich.iteration.service;
 import com.poortorich.accountbook.entity.AccountBook;
 import com.poortorich.accountbook.enums.AccountBookType;
 import com.poortorich.accountbook.request.AccountBookRequest;
-import com.poortorich.accountbook.response.AccountBookResponse;
+import com.poortorich.accountbook.response.enums.AccountBookResponse;
 import com.poortorich.accountbook.util.AccountBookBuilder;
 import com.poortorich.accountbook.entity.enums.IterationType;
 import com.poortorich.accountbook.request.enums.IterationAction;
@@ -18,7 +18,7 @@ import com.poortorich.iteration.repository.IterationRepository;
 import com.poortorich.iteration.request.CustomIteration;
 import com.poortorich.iteration.request.IterationRule;
 import com.poortorich.iteration.response.CustomIterationInfoResponse;
-import com.poortorich.iteration.response.IterationResponse;
+import com.poortorich.iteration.response.enums.IterationResponse;
 import com.poortorich.iteration.util.IterationBuilder;
 import com.poortorich.user.entity.User;
 import lombok.RequiredArgsConstructor;
